Timer text bg, formatted slice duration
Added semi-transparent background on timer text and resized the pie
chart to fill the available space. Fixed the problem with preset colors
on the pie chart not matching the tasks in the bottom drawer, but now
the colors start on purple instead of red for new tasks. Seconds now
appear on the pie chart slices and on the slice duration. Slice duration
formatted to be HH:MM:SS. Implemented more CustomColors.